+++ eq.go
@@ -412,7 +412,7 @@ func (r *eqRule[T, V]) search(v T, dst *roaring.Bitmap, pool *bitmapPool) {
 		return
 	}
 
-	bits := cache.replace(value)
+	bits := cache.replace(value, pool)
 	r.addMatches(value, bits)
 	dst.Or(bits)
 }
@@ -530,7 +530,7 @@ func (r *unaryEqRule[T, V]) search(v T, dst *roaring.Bitmap, pool *bitmapPool) {
 		r.addMatches(value, dst)
 		return
 	}
-	bits := cache.replace(value)
+	bits := cache.replace(value, pool)
 	r.addMatches(value, bits)
 	dst.Or(bits)
 }
@@ -621,7 +621,7 @@ func (r *binaryEqRule[T, V]) search(v T, dst *roaring.Bitmap, pool *bitmapPool) 
 		r.addMatches(value, dst)
 		return
 	}
-	bits := cache.replace(value)
+	bits := cache.replace(value, pool)
 	r.addMatches(value, bits)
 	dst.Or(bits)
 }