+++ not.go
@@ -65,7 +65,7 @@ func (r *notRule[T, V]) exclude(v T, dst *roaring.Bitmap, pool *bitmapPool) {
 		return
 	}
 
-	bits := cache.replace(value)
+	bits := cache.replace(value, pool)
 	r.addExclusions(value, bits)
 	dst.Or(bits)
 }